修警告

+++ kernel/src/gdt.c
@@ -33,7 +33,7 @@ static inline void lgdt(struct pseudodesc *pd) {
 unsigned char stack0[1024];
 
 void gdt_init() {
-  ts.esp0 = stack0 + sizeof(stack0);
+  ts.esp0 = (uint32_t)stack0 + sizeof(stack0);
   ts.ss0 = KERNEL_DS;
 
   // initialize the TSS filed of the gdt
